[sms] Unable to receive SMS messages
The problem is that Qtopia does not retry long enough to send
the AT+CNMI message before giving up.  (This message instructs
the GSM to send an unsolicited message upon arrival of an SMS.)

Notes in the code indicate that some modems will be busy for a
while after the AT+CPIN transaction, and Qtopia retries 5 times,
each time one second apart.

However, the docs for the Enfora Enabler, which have been noted
to be more similar to the Calypso than others, make a specific
note that it may take 20 seconds or more after power-up before
the SIM is read, and certain SIM-related commmands can be run.
(The error we get from the Calypso is "SIM BUSY".)  Various tests
here confirm that the magic number for the Calypso with the SIM
in question (a T-Mobile SIM) is 20 seconds.

The patch below changes the 5 retries to 25 -- which solves my SMS
problems completely, and is large enough that it should work for
others as well.  There is no impact for SIMs that worked before.